Move model stubs to separate package to be shared among tests
