+++ src/test/java/seedu/address/testutil/modelstubs/ModelStub.java
@@ -1,0 +1,67 @@
+package seedu.address.testutil.modelstubs;
+
+import static org.junit.Assert.fail;
+
+import java.util.Collection;
+import java.util.function.Predicate;
+
+import javafx.collections.ObservableList;
+import seedu.address.model.Model;
+import seedu.address.model.ReadOnlyAddressBook;
+import seedu.address.model.person.ReadOnlyPerson;
+import seedu.address.model.person.exceptions.DuplicatePersonException;
+import seedu.address.model.person.exceptions.PersonNotFoundException;
+import seedu.address.model.tag.Tag;
+
+/**
+ * A default model stub that have all of the methods failing.
+ */
+public class ModelStub implements Model {
+    @Override
+    public void addPerson(ReadOnlyPerson person) throws DuplicatePersonException {
+        fail("This method should not be called.");
+    }
+
+    @Override
+    public void addPersons(Collection<ReadOnlyPerson> persons) {
+        fail("This method should not be called.");
+    }
+
+    @Override
+    public void resetData(ReadOnlyAddressBook newData) {
+        fail("This method should not be called.");
+    }
+
+    @Override
+    public ReadOnlyAddressBook getAddressBook() {
+        fail("This method should not be called.");
+        return null;
+    }
+
+    @Override
+    public void deletePerson(ReadOnlyPerson target) throws PersonNotFoundException {
+        fail("This method should not be called.");
+    }
+
+    @Override
+    public void updatePerson(ReadOnlyPerson target, ReadOnlyPerson editedPerson)
+            throws DuplicatePersonException {
+        fail("This method should not be called.");
+    }
+
+    @Override
+    public ObservableList<ReadOnlyPerson> getFilteredPersonList() {
+        fail("This method should not be called.");
+        return null;
+    }
+
+    @Override
+    public void updateFilteredPersonList(Predicate<ReadOnlyPerson> predicate) {
+        fail("This method should not be called.");
+    }
+
+    @Override
+    public void removeTag(Tag tag) throws PersonNotFoundException, DuplicatePersonException {
+        fail("This method should not be called.");
+    }
+}
